clustectl integration

+++ Makefile
@@ -110,14 +110,24 @@ docker-push: ## Push docker image with the manager.
 
 ##@ Deployment
 
+set_controller_image:
+	cd config/manager && $(KUSTOMIZE) edit set image controller=${IMG}
+
+create_infrastructure-components: manifests kustomize
+ifeq ($(IMG),controller:latest)
+	echo "Please set IMG env var"
+	exit 1
+endif
+	make set_controller_image
+	$(KUSTOMIZE) build config/default > infrastructure-components.yaml
+
 install: manifests kustomize ## Install CRDs into the K8s cluster specified in ~/.kube/config.
 	$(KUSTOMIZE) build config/crd | kubectl apply -f -
 
 uninstall: manifests kustomize ## Uninstall CRDs from the K8s cluster specified in ~/.kube/config.
 	$(KUSTOMIZE) build config/crd | kubectl delete -f -
 
-deploy: manifests kustomize ## Deploy controller to the K8s cluster specified in ~/.kube/config.
-	cd config/manager && $(KUSTOMIZE) edit set image controller=${IMG}
+deploy: manifests kustomize set_controller_image ## Deploy controller to the K8s cluster specified in ~/.kube/config.
 	$(KUSTOMIZE) build config/default | kubectl apply -f -
 
 undeploy: ## Undeploy controller from the K8s cluster specified in ~/.kube/config.
